requirements.txt
blessings==1.6
bpython==0.16
coreapi==2.1.1
curtsies==0.2.11
Django==1.10.3
django-admin-sortable==2.0.21
django-autoslug==1.9.3
django-bootstrap3==7.1.0
django-extensions==1.7.4
django-markdownx==1.6.3
django-rest-swagger==2.1.1
django-reversion==2.0.7
djangorestframework==3.5.3
drf-nested-routers==0.11.1
greenlet==0.4.10
itypes==1.1.0
Markdown==2.6.7
openapi-codec==1.2.1
Pillow==3.4.2
Pygments==2.1.3
requests==2.11.1
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
wcwidth==0.1.7
Werkzeug==0.11.11
